Fix Bug on Reporting View

+++ mepa-core/src/main/java/fr/epita/sigl/mepa/core/dao/GameDao.java
@@ -16,8 +16,8 @@ public interface GameDao {
 
     List<Game> getAll();
 
-	int getComingGameById(Long id);
+	Long getComingGameById(Long id);
 
-	int getEndedGameById(Long id);
+	Long getEndedGameById(Long id);
 
 }
